add two class for configuration url - with env.properties and testng.xml
