+++ docker-compose.yml
@@ -1,0 +1,4 @@
+FROM ubuntu:latest
+LABEL authors="Даниил"
+
+ENTRYPOINT ["top", "-b"]
